Move announce send to tokio tasks
