+++ dds/src/dds/domain/domain_participant.rs
@@ -5,9 +5,14 @@ use tokio::sync::mpsc::{Receiver, Sender};
 use crate::{
     builtin_topics::{ParticipantBuiltinTopicData, TopicBuiltinTopicData},
     implementation::{
+        data_representation_builtin_endpoints::{
+            discovered_reader_data::{DiscoveredReaderData, ReaderProxy},
+            discovered_topic_data::DiscoveredTopicData,
+            discovered_writer_data::{DiscoveredWriterData, WriterProxy},
+        },
         dds::{
             any_topic_listener::AnyTopicListener,
-            dds_domain_participant::DdsDomainParticipant,
+            dds_domain_participant::{AnnounceKind, DdsDomainParticipant},
             nodes::{
                 DataReaderNode, DataWriterNode, DomainParticipantNode, SubscriberNode,
                 SubscriberNodeKind, TopicNode, TopicNodeKind,
@@ -32,7 +37,11 @@ use crate::{
     },
     publication::{publisher::Publisher, publisher_listener::PublisherListener},
     subscription::{subscriber::Subscriber, subscriber_listener::SubscriberListener},
-    topic_definition::{topic::Topic, topic_listener::TopicListener, type_support::DdsType},
+    topic_definition::{
+        topic::Topic,
+        topic_listener::TopicListener,
+        type_support::{dds_serialize, DdsSerializedKey, DdsType},
+    },
 };
 
 use super::{
@@ -659,6 +668,42 @@ impl DomainParticipant {
 }
 
 /////////////////////////////////////////////////////////////////////////
+
+pub async fn task_send_entity_announce(
+    participant_guid_prefix: GuidPrefix,
+    mut announce_receiver: Receiver<AnnounceKind>,
+) {
+    loop {
+        if let Some(announce_kind) = announce_receiver.recv().await {
+            THE_DDS_DOMAIN_PARTICIPANT_FACTORY.get_participant_mut(&participant_guid_prefix, |dp| {
+                if let Some(dp) = dp {
+                    announce_entity(dp, announce_kind);
+                }
+            })
+        }
+    }
+}
+
+fn announce_entity(domain_participant: &mut DdsDomainParticipant, announce_kind: AnnounceKind) {
+    match announce_kind {
+        AnnounceKind::CreatedDataReader(discovered_reader_data) => {
+            announce_created_data_reader(domain_participant, discovered_reader_data)
+        }
+        AnnounceKind::CreatedDataWriter(discovered_writer_data) => {
+            announce_created_data_writer(domain_participant, discovered_writer_data)
+        }
+        AnnounceKind::CratedTopic(discovered_topic_data) => {
+            announce_created_topic(domain_participant, discovered_topic_data)
+        }
+        AnnounceKind::DeletedDataReader(deleted_reader_handle) => {
+            announce_deleted_reader(domain_participant, deleted_reader_handle)
+        }
+        AnnounceKind::DeletedDataWriter(deleted_writer_handle) => {
+            announce_deleted_writer(domain_participant, deleted_writer_handle)
+        }
+        AnnounceKind::DeletedParticipant => (),
+    }
+}
 
 pub async fn task_user_defined_receive(
     participant_guid_prefix: GuidPrefix,
@@ -1413,4 +1458,147 @@ fn on_inconsistent_topic_communication_change(topic_node: TopicNode) {
             l.add_communication_state(status_kind);
         }
     })
+}
+
+fn announce_created_data_reader(
+    domain_participant: &mut DdsDomainParticipant,
+    discovered_reader_data: DiscoveredReaderData,
+) {
+    let reader_proxy = ReaderProxy::new(
+        discovered_reader_data.reader_proxy().remote_reader_guid(),
+        discovered_reader_data
+            .reader_proxy()
+            .remote_group_entity_id(),
+        domain_participant.default_unicast_locator_list().to_vec(),
+        domain_participant.default_multicast_locator_list().to_vec(),
+        discovered_reader_data.reader_proxy().expects_inline_qos(),
+    );
+    let reader_data = &DiscoveredReaderData::new(
+        reader_proxy,
+        discovered_reader_data
+            .subscription_builtin_topic_data()
+            .clone(),
+    );
+
+    let serialized_data = dds_serialize(reader_data).expect("Failed to serialize data");
+
+    let timestamp = domain_participant.get_current_time();
+    domain_participant
+        .get_builtin_publisher_mut()
+        .stateful_data_writer_list_mut()
+        .iter_mut()
+        .find(|x| x.get_type_name() == DiscoveredReaderData::type_name())
+        .unwrap()
+        .write_w_timestamp(
+            serialized_data,
+            reader_data.get_serialized_key(),
+            None,
+            timestamp,
+        )
+        .expect("Should not fail to write built-in message");
+}
+
+fn announce_created_data_writer(
+    domain_participant: &mut DdsDomainParticipant,
+    discovered_writer_data: DiscoveredWriterData,
+) {
+    let writer_data = &DiscoveredWriterData::new(
+        discovered_writer_data.dds_publication_data().clone(),
+        WriterProxy::new(
+            discovered_writer_data.writer_proxy().remote_writer_guid(),
+            discovered_writer_data
+                .writer_proxy()
+                .remote_group_entity_id(),
+            domain_participant.default_unicast_locator_list().to_vec(),
+            domain_participant.default_multicast_locator_list().to_vec(),
+            discovered_writer_data
+                .writer_proxy()
+                .data_max_size_serialized(),
+        ),
+    );
+
+    let serialized_data = dds_serialize(writer_data).expect("Failed to serialize data");
+
+    let timestamp = domain_participant.get_current_time();
+
+    domain_participant
+        .get_builtin_publisher_mut()
+        .stateful_data_writer_list_mut()
+        .iter_mut()
+        .find(|x| x.get_type_name() == DiscoveredWriterData::type_name())
+        .unwrap()
+        .write_w_timestamp(
+            serialized_data,
+            writer_data.get_serialized_key(),
+            None,
+            timestamp,
+        )
+        .expect("Should not fail to write built-in message");
+}
+
+fn announce_created_topic(
+    domain_participant: &mut DdsDomainParticipant,
+    discovered_topic: DiscoveredTopicData,
+) {
+    let serialized_data = dds_serialize(&discovered_topic).expect("Failed to serialize data");
+
+    let timestamp = domain_participant.get_current_time();
+
+    domain_participant
+        .get_builtin_publisher_mut()
+        .stateful_data_writer_list_mut()
+        .iter_mut()
+        .find(|x| x.get_type_name() == DiscoveredTopicData::type_name())
+        .unwrap()
+        .write_w_timestamp(
+            serialized_data,
+            discovered_topic.get_serialized_key(),
+            None,
+            timestamp,
+        )
+        .expect("Should not fail to write built-in message");
+}
+
+fn announce_deleted_reader(
+    domain_participant: &mut DdsDomainParticipant,
+    reader_handle: InstanceHandle,
+) {
+    let serialized_key = DdsSerializedKey::from(reader_handle.as_ref());
+    let instance_serialized_key =
+        cdr::serialize::<_, _, cdr::CdrLe>(&serialized_key, cdr::Infinite)
+            .map_err(|e| DdsError::PreconditionNotMet(e.to_string()))
+            .expect("Failed to serialize data");
+
+    let timestamp = domain_participant.get_current_time();
+
+    domain_participant
+        .get_builtin_publisher_mut()
+        .stateful_data_writer_list_mut()
+        .iter_mut()
+        .find(|x| x.get_type_name() == DiscoveredReaderData::type_name())
+        .unwrap()
+        .dispose_w_timestamp(instance_serialized_key, reader_handle, timestamp)
+        .expect("Should not fail to write built-in message");
+}
+
+fn announce_deleted_writer(
+    domain_participant: &mut DdsDomainParticipant,
+    writer_handle: InstanceHandle,
+) {
+    let serialized_key = DdsSerializedKey::from(writer_handle.as_ref());
+    let instance_serialized_key =
+        cdr::serialize::<_, _, cdr::CdrLe>(&serialized_key, cdr::Infinite)
+            .map_err(|e| DdsError::PreconditionNotMet(e.to_string()))
+            .expect("Failed to serialize data");
+
+    let timestamp = domain_participant.get_current_time();
+
+    domain_participant
+        .get_builtin_publisher_mut()
+        .stateful_data_writer_list_mut()
+        .iter_mut()
+        .find(|x| x.get_type_name() == DiscoveredWriterData::type_name())
+        .unwrap()
+        .dispose_w_timestamp(instance_serialized_key, writer_handle, timestamp)
+        .expect("Should not fail to write built-in message");
 }
